refactor: move `compute_lagrange_interpolated_poly` tests to `lagrange_poly.rs`
